Change package configuration
This creates a suitable entry point and move the barcodex code to the root
level so that it can be imported by `import barcodex` rather than `import
barcodex.barcodex`.
