+++ setup.py
@@ -16,35 +16,33 @@ with open("README.md") as infile:
     content = infile.read().rstrip()
 
 setup(
-	name = "barcodex",
-	version = __version__,
-	author = "Richard Jovelin",
-	author_email = "richard.jovelin@oicr.on.ca",
-	description = ("A tool for extracting Unique Molecular Identifiers (UMIs) \
+    name="barcodex",
+    version=__version__,
+    author="Richard Jovelin",
+    author_email="richard.jovelin@oicr.on.ca",
+    description=("A tool for extracting Unique Molecular Identifiers (UMIs) \
                 from single or paired-end read sequences"),
-	license = "MIT License",
-	keywords = "computational genomics",
-	url = "https://github.com/oicr-gsi/barcodex",
-	packages = ['barcodex'],
-	long_description = content,
-	long_description_content_type='text/markdown',
-    classifiers = [
-	"Development Status :: 3 - Alpha",
-	"Intended Audience :: Science/Research",
-	"Intended Audience :: Developers",
-	"License :: OSI Approved :: MIT License",
-	"Programming Language :: Python :: 3",
-	"Programming Language :: Python :: 3.6",
-	"Topic :: Software Development",
-	"Topic :: Scientific/Engineering",
-	"Operating System :: POSIX",
-	"Operating System :: Unix",
-	"Operating System :: MacOS",
-    "Operating System :: Microsoft :: Windows"
+    license="MIT License",
+    keywords="computational genomics",
+    url="https://github.com/oicr-gsi/barcodex",
+    packages=['barcodex'],
+    long_description=content,
+    long_description_content_type='text/markdown',
+    classifiers=[
+        "Development Status :: 3 - Alpha",
+        "Intended Audience :: Science/Research",
+        "Intended Audience :: Developers",
+        "License :: OSI Approved :: MIT License",
+        "Programming Language :: Python :: 3",
+        "Programming Language :: Python :: 3.6",
+        "Topic :: Software Development",
+        "Topic :: Scientific/Engineering",
+        "Operating System :: POSIX",
+        "Operating System :: Unix",
+        "Operating System :: MacOS",
+        "Operating System :: Microsoft :: Windows"
     ],
-    entry_points={'console_scripts': ['barcodex = barcodex.py:main']},
-    install_requires = ["regex>=2020.6.8"],
+    entry_points={'console_scripts': ['barcodex = barcodex:main']},
+    install_requires=["regex>=2020.6.8"],
     python_requires=">=3.6",
 )
-    
-
